requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
config==0.5.0.post0
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.0
humanfriendly==8.1
idna==2.9
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
PyDispatcher==2.0.5
PyGObject==3.36.0
PyGObject-stubs==0.0.2
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
stopwatch.py==1.0.1
treelib==1.6.1
uritemplate==3.0.1
urllib3==1.25.8
